Prevent occasional launch flaw
Sometimes the note would be barely touching the thrusters.  Since the thrusters start basically at the same time the note is pulled backward, about 10% of the time the note would shoot out of the launcher and land a mild metre or two from the front of the robot.  Now there is a delay of 0.1 seconds just so that the note has some time to escape the clutches of the thrusters before they start.