Stop thrusters from spinning if intaking
Immediately after launching, the thrusters are still spinning.  We cannot put a brake on them because they spin so fast.  When we took in a note while this was happening, sometimes it would go right through the robot and plop onto the ground in front of it.  These additional four lines of code should stop this from happening.