Add automatic pickup to raw mode
Previously, we thought that the automatic pickup should not be added to raw mode so that we could have a fallback option in case of a sensor failure.  We have, however, over time come to realise that we must be dependent on these sensors anyway and that they are worth depending upon.  Raw mode is mainly for the case of a camera or navx failure, not a key sensor failure.