Add sensor error protection
If there is a sensor error happening, the robot will still be able to collect and score notes, though not nearly as easily or as well.  If a sensor is broken during a match, all the driver or co-driver has to do is turn the on/off switch in Shuffleboard for "Sensor Error" to on.  The robot will better adhere to manual control and understand that its inputs are delusional and not to be trusted.