If speaker not in view in side auto, rotate until
When a side auto picks up a note and prepares to score it, it must point towards the speaker as a component of aiming.  What if the speaker is not in view?  Now it will rotate until it sees the speaker.  This is a safe thing to add to the code, especially given that it does not affect our solid centre auto.