Solve LAUNCHprep raw mode problem
Whenever we used LAUNCHprep (right button) to try to score a note, it would never actually launch the note.  This, I realise now, is because we were calling the wrong function: the function that needs the camera to aim and prepare to launch.  This is certainly a safe change to make to the code since it will fix something that wasn't working in the first place.