Move amp-scoring functionality out of zero zone
The "zero zone" in the code is the stuff in the if statement dependent upon `launcher.stage == 0`.  We had some problems with this during matches because the robot would be trying to launch or intake something when we needed to override that and go for the amp.  This is a safe change to make to the code.