EZ way to score while pressed up against speaker
Now, pressing and holding the left button until a launch is over will simply launch a note.  Now, one click (not holding down) of the left button will launch after going to the closeup aim encoder position.  This is clearly a safe and useful change to make to the code.