Faster Amp Launch
When Y is pressed on the secondary controller, the robot will rotate to face the amp AND point the launcher down AND pull the note back AND start the thrusters.  When the Y button is released, the launcher will fire the note into the amp.  This saves us approximately 2.5 seconds per amp launch.